+++ upload/api/language/ru-ru/logs/messages.php
@@ -3,4 +3,5 @@ return [
     'request_list' => 'Логи запросов',
     'security_list' => 'Логи безопасности',
     'audit_list' => 'Логи аудита',
+    'frontend_error_chart' => 'Ошибки фронтенда по часам',
 ];